+++ R/ColumnLinkedMatrix.R
@@ -217,6 +217,10 @@ setMethod("initialize", "ColumnLinkedMatrix", function(.Object, ...) {
     if (any(unlist(lapply(list, function(x) length(dim(x)) != 2)))) {
         stop("arguments need to be matrix-like")
     }
+    # Detect matrices that do not match in dimensions
+    if (ncol(unique(sapply(list, dim), MARGIN = 2)) != 1) {
+        stop("arguments need the same number of rows")
+    }
     .Object <- callNextMethod(.Object, list)
     return(.Object)
 })