+++ NEWS.md
@@ -1,7 +1,9 @@
 # LinkedMatrix 1.0.0.9000
 
 * Remove `apply` and derivative functions to keep the package minimal.
-* Only allow matrix-like objects when creating linked matrices.
+* Only allow matrix-like objects with matching row or column dimensions (same
+  number of rows in case of `ColumnLinkedMatrix` and same number of columns in
+  case of `RowLinkedMatrix`) when creating linked matrices.
 
 # LinkedMatrix 1.0.0